split viewer test
